Fix cypress EOI tests (#642)
* Add BASE_URL environment variable to configure which environment cypress is targeting for testing

* Add option for cypress to use HTTP Basic authentication when running tests

* Identify cypress tests with a special user agent

* Fix gov.uk caption selector

* Fix Gov.UK guidance page heading selectors

* Fix DoB of someone who is exactly 18 years old

This test was previously using 31/09/2007 as someone who was over 18 on 01/09/2025, when in fact this DoB makes the person 17.

* Fix EOI links test header selector

* Migrate cypress tests to using cy.session instead of Cypress.Cookies

* Fix EOI address cypress test

Make each test not dependent on implicit state by starting a new session before each test. This means we had to fix a lot of element selectors to use the regular input box instead of the error state one

* Make more EOI cypress tests independent of previous tests running

* E2E start date tests passing

* Fix EOI error label cypress tests

* Remove erroneous test focus

* Legacy journey updated

* Fix EOI links cypress test

Tests that target page contents on different origins needs to run in it's own `cy.origin`

* Use a new state each time on the EOI phone number cypress test

* Use a fresh state each time for the EOI postcode cypress tests

* Fix EOI self-assessment unhappy path test

---------

Co-authored-by: Will Gibson <8738245+WillGibson@users.noreply.github.com>